refactor(ts): extract timeline/persistence.ts from store
Move the durable-state boundary (localStorage autosave, project file import/export,
client-side gzip share codes, getProjectData/loadProjectData) into a
useTimelinePersistence(deps) composable, mirroring the optimizer. Function bodies
moved verbatim; only store bindings, the shared snapshot core (_loadSnapshot,
restoreArmoryFromSnapshot), and armory stores are injected, so the load-bearing
serialization schema (STORAGE_KEY, snapshot shape) is unchanged. Scenario CRUD
stays inline for now. No behavior change.
